Move cost filtering into Advanced Filters
