+++ src/app/pages/partners/analytics-events.js
@@ -1,5 +1,5 @@
 import analytics from '~/helpers/analytics';
-import {books, costs, types, advanced} from './store';
+import {books, types, advanced} from './store';
 
 function toLabel(store) {
     return store.value.join(',') || 'N/A';
@@ -18,12 +18,11 @@ function sendAddEvent(actionObj) {
 }
 
 function filterIsSelected() {
-    return types.value || advanced.value.length > 0 || costs.value;
+    return types.value || advanced.value.length > 0;
 }
 
 types.on('notify', sendAddEvent);
 advanced.on('notify', sendAddEvent);
-costs.on('notify', sendAddEvent);
 books.on('notify', (obj) => {
     const addingBook = typeof obj === 'object' && 'add' in obj;
